+++ src/main/java/shadows/apotheosis/adventure/affix/effect/ThunderstruckAffix.java
@@ -60,7 +60,7 @@ public class ThunderstruckAffix extends Affix {
 		if (Apotheosis.localAtkStrength >= 0.98) {
 			List<Entity> nearby = target.level.getEntities(target, new AABB(target.blockPosition()).inflate(6), CleavingAffix.cleavePredicate(user, target));
 			for (Entity e : nearby) {
-				e.hurt(DamageSource.LIGHTNING_BOLT, getTrueLevel(rarity, level));
+				e.hurt(DamageSource.mobAttack(user), getTrueLevel(rarity, level));
 			}
 		}
 	}